Lucene's snowball changes
* Add armenian stemmer: https://snowballstem.org/algorithms/armenian/stemmer.html
* Add estonian stemmer: https://github.com/snowballstem/snowball/pull/108
* Enable german2, kraaij_pohlman, and lovins stemmers
* fix generator to use MethodHandle.invokeExact instead of super-slow reflection
* remove String object creation overhead from stemmer
* add missing javadocs to public classes

there is a script (lucene.sh) to regenerate the classes and populate the
lucene source tree with the new ones. It takes care of a few things that
aren't cleanly fixed yet:

* unconventional classnames (frenchStemmer -> FrenchStemmer)
* mixed tab/space indendentation
* missing license headers

This isn't enough to guarantee lucene tests will pass. For example some
stemmers tests simply check expected vs. actual stemming of results from a
dictionary. If the stemmers have been modified, we need to regenerate
test data too.